third_party/WebKit/LayoutTests/svg/hixie/transform/001-expected.txt

CONSOLE ERROR: line 3: Error: <circle> attribute transform: Unexpected end of attribute. Expected ')', "translate(200,0".
